re-structuring and genericising of cancellation flow and few accompanying wider improvements/refactors
