+++ app/client/components/cancel/cancellationReason.ts
@@ -1,0 +1,13 @@
+export interface CancellationReason {
+  reasonId: string;
+  linkLabel: string;
+  saveTitle: string;
+  saveBody: string | JSX.Element;
+  experimentSaveBody?: JSX.Element;
+  experimentTriggerFlag?: string;
+  alternateCallUsPrefix?: string;
+  alternateFeedbackIntro?: string;
+  alternateFeedbackThankYouTitle?: string;
+  alternateFeedbackThankYouBody?: string;
+  skipFeedback?: boolean;
+}
