+++ app/client/components/cancel/membership/membershipCancellationReasons.tsx
@@ -1,7 +1,8 @@
 import React from "react";
-import { conf } from "../../../server/config";
-import palette from "../../colours";
-import { SwitchToContributionPlaceholder } from "./membership/switchToContributionPlaceholder";
+import { conf } from "../../../../server/config";
+import palette from "../../../colours";
+import { CancellationReason } from "../cancellationReason";
+import { SwitchToContributionPlaceholder } from "./switchToContributionPlaceholder";
 
 // Webpack doesn't like browser globals
 let domain: string;
@@ -11,21 +12,7 @@ if (typeof window !== "undefined" && window.guardian) {
   domain = conf.DOMAIN;
 }
 
-export interface CancellationReason {
-  reasonId: string;
-  linkLabel: string;
-  saveTitle: string;
-  saveBody: string | JSX.Element;
-  experimentSaveBody?: JSX.Element;
-  experimentTriggerFlag?: string;
-  alternateCallUsPrefix?: string;
-  alternateFeedbackIntro?: string;
-  alternateFeedbackThankYouTitle?: string;
-  alternateFeedbackThankYouBody?: string;
-  skipFeedback?: boolean;
-}
-
-export const membershipCancellationReasonMatrix: CancellationReason[] = [
+export const membershipCancellationReasons: CancellationReason[] = [
   {
     reasonId: "mma_financial_circumstances",
     linkLabel: "A change in my financial circumstances",
